align module name
